requirements.txt
nextcord==2.0.0a3
koreanbots==1.1.1
requests==2.26.0
xmltodict==0.12.0
asyncio==3.4.3
python-dotenv==0.19.0
wheel==0.37.0
jwt==1.3.0